requirements.txt
psycopg2-binary==2.9.7
tabulate==0.9.0
colorama==0.4.6